+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Please avoid adding duplicate information across this changelog and JIRA/doc inp
 
 - Telemetry collection enabled by default for self-managed installations. Opt-out is available on bundle generation, or at any time via the System Configuration UI.
 - A new environment variable `ROX_REGISTRY_MIRRORING_ENABLED` has been added that when set to `true` will enable processing registry mirrors during image enrichment. Mirror details are obtained via the `ImageContentSourcePolicy`, `ImageDigestMirrorSet`, and `ImageTagMirrorSet` CRs.
-- ROX-17702: Usage metrics collection API: `/v1/usage`, `/v1/usage/max`, `/v1/usage/current`. New `/api/usage/csv` endpoint.
+- ROX-17702: Usage metrics collection API: `/v1/usage/current`, `/v1/usage/max`. New `/api/usage/csv` endpoint.
 
 ### Removed Features